feat: create enum file for attendee-type
