+++ api/v1/model/enum.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class AttendeeType(str, Enum):
+    vip = "vip"
+    worker = "worker"
+    normal-guest = "normal-guest"
