+++ api/v1/model/user.py
@@ -1,13 +1,11 @@
-from sqlalchemy import Column, Integer, String
-from sqlalchemy.orm import relationship
-from db.database import Base
+# from sqlalchemy import Column, Integer, String
+# from sqlalchemy.orm import relationship
+# from db.database import Base
 
-class User(Base):
-    __tablename__ = "users"
+# class User(Base):
+#     __tablename__ = "users"
 
-    id = Column(Integer, primary_key=True, index=True)
-    username = Column(String(50), unique=True, index=True)
-    email = Column(String(100), unique=True, index=True)
-    password = Column(String(255))
+#     id = Column(Integer, primary_key=True, index=True)
+#     email = Column(String(100), unique=True, index=True)
+#     password = Column(String(255)) 
 
-
